Remove hardcoded value for HTTP code

+++ src/itti/msgs/itti_msg_sbi.hpp
@@ -258,7 +258,7 @@ class itti_sbi_register_nf_instance_response : public itti_sbi_msg {
   const char* get_msg_name() { return "SBI_REGISTER_NF_INSTANCE_RESPONSE"; };
 
   amf_application::amf_profile profile;
-  uint8_t http_response_code;
+  uint16_t http_response_code;
   std::string nrf_uri;
 };
